requirements.txt
customtkinter==5.0.3
Pillow==9.0.1
matplotlib==3.5.1
